refactor: search page 및 실내 바텀시트 섹션 분리(#49)
